remove catalyst.jax_primitives.subroutine
